+++ com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/model/TestKind.java
@@ -1,5 +1,5 @@
 package com.microsoft.java.test.plugin.model;
 
 public enum TestKind {
-    JUnit, JUnit5,
+    JUnit, JUnit5, TestNG
 }
