+++ com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/searcher/JUnit5TestSearcher.java
@@ -12,7 +12,7 @@
 package com.microsoft.java.test.plugin.searcher;
 
 import com.microsoft.java.test.plugin.model.TestKind;
-import com.microsoft.java.test.plugin.util.JUnitUtility;
+import com.microsoft.java.test.plugin.util.TestSearchUtils;
 
 import org.eclipse.jdt.core.Flags;
 import org.eclipse.jdt.core.IMethod;
@@ -20,7 +20,7 @@ import org.eclipse.jdt.core.JavaModelException;
 
 public class JUnit5TestSearcher implements TestFrameworkSearcher {
 
-    public static final String JUNIT_TEST_ANNOTATION = "org.junit.jupiter.api.Test";
+    public static final String TEST_METHOD_ANNOTATION = "org.junit.jupiter.api.Test";
 
     @Override
     public TestKind getTestKind() {
@@ -35,7 +35,7 @@ public class JUnit5TestSearcher implements TestFrameworkSearcher {
             // 'V' is void signature
             return !(method.isConstructor() || Flags.isAbstract(flags) ||
                     Flags.isStatic(flags) || !"V".equals(method.getReturnType())) &&
-                    JUnitUtility.hasTestAnnotation(method, JUNIT_TEST_ANNOTATION);
+                    TestSearchUtils.hasTestAnnotation(method, TEST_METHOD_ANNOTATION);
         } catch (final JavaModelException e) {
             // ignore
             return false;
