--- com.microsoft.java.test.plugin/src/main/java/com/microsoft/java/test/plugin/util/JUnitUtility.java
@@ -1,69 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2017 Microsoft Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     Microsoft Corporation - initial API and implementation
- *******************************************************************************/
-
-package com.microsoft.java.test.plugin.util;
-
-import org.eclipse.jdt.core.Flags;
-import org.eclipse.jdt.core.IAnnotation;
-import org.eclipse.jdt.core.IClassFile;
-import org.eclipse.jdt.core.ICompilationUnit;
-import org.eclipse.jdt.core.IJavaElement;
-import org.eclipse.jdt.core.IMethod;
-import org.eclipse.jdt.core.IType;
-import org.eclipse.jdt.core.JavaModelException;
-
-import java.util.Arrays;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-public class JUnitUtility {
-    public static boolean isAccessibleClass(IType type) throws JavaModelException {
-        int flags = type.getFlags();
-        if (Flags.isInterface(flags)) {
-            return false;
-        }
-        IJavaElement parent = type.getParent();
-        while (true) {
-            if (parent instanceof ICompilationUnit || parent instanceof IClassFile) {
-                return true;
-            }
-            if (!(parent instanceof IType) || !Flags.isStatic(flags) || !Flags.isPublic(flags)) {
-                return false;
-            }
-            flags = ((IType) parent).getFlags();
-            parent = parent.getParent();
-        }
-    }
-
-    public static boolean hasTestAnnotation(IMethod method, String annotation) {
-        try {
-            final Optional<IAnnotation> matched = Arrays.stream(method.getAnnotations())
-                    .filter(a -> annotation.endsWith(a.getElementName())).findAny();
-            if (!matched.isPresent()) {
-                return false;
-            }
-            final IAnnotation anno = matched.get();
-            if (!anno.exists()) {
-                return false;
-            }
-            final String name = anno.getElementName();
-            final String[][] fullNameArr = method.getDeclaringType().resolveType(name);
-            if (fullNameArr == null) {
-                final ICompilationUnit cu = method.getCompilationUnit();
-                return cu != null && cu.getImport(annotation).exists();
-            }
-            final String fullName = Arrays.stream(fullNameArr[0]).collect(Collectors.joining("."));
-            return fullName.equals(annotation);
-        } catch (final JavaModelException e) {
-            return false;
-        }
-    }
-}
